% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{addcntcat}
\alias{addcntcat}
\title{addcntcat adds the years of experience by record to data.frame}
\usage{
addcntcat(
  indat,
  diver = "diver",
  year = "year",
  catch = "catch",
  count = "count",
  totC = "totC"
)
}
\arguments{
\item{indat}{the docket catch-effort data.frame of interest}

\item{diver}{character string name of the active diver/vessel or whoever/
whatever is fishing each year, default='diver'}

\item{year}{character string name of the year, default='year'}

\item{catch}{character string name of the catch variable in the data.frame}

\item{count}{character string name for the count variable to be added to the 
data.frame, default='count'}

\item{totC}{character string name for the totalC variable to be added to the 
data.frame, default='totC'}
}
\value{
the same input data.frame with the addition of two columns
}
\description{
addcntcat adds the count of years of experience, and the total
    catch across years, for each diver, as recorded in the input catch-effort 
    data.frame. Additions are made to the right-hand side of the input 
    data.frame. The count of years and total catch for each diver will 
    obviously change depending on which block, or collection of blocks, are 
    selected. The value of this is that it facilitates the exploration of the 
    influence of years of experience and total catch on everything else.
}
\examples{
\dontrun{
  print("waiting on an integral data-set")
}
}
