% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{addcount}
\alias{addcount}
\title{addcount adds a count of years and avC to a data.frame}
\usage{
addcount(indat, varid = "diver", group = "year", catch = "catch")
}
\arguments{
\item{indat}{a data.frame containing at least columns named 'catch',
'diver' (or in fact 'varid'), and 'year'.}

\item{varid}{the variable in the data.frame for which the count and
average catch per year is to be estimated. Defaults to 'diver'
ready for abalone data but could be vessel for other fisheries,
or any other factor of interest.}

\item{group}{the first grouping variable, typically diver ifor abalone}

\item{catch}{the variable name for the catch, sometimes blip in abalone}
}
\value{
a data.frame made of the input data.frame plus two new columns
}
\description{
addcount given an input data.frame containing catch,
    diver, and year as variables this function first sums the catch by
    diver, by year, then counts each diver's occurrence across the
    years, and each diver's catch. Then stepping through each diver it
    populates two new colummns named 'count' and 'avC'.
}
