% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{addyrexp}
\alias{addyrexp}
\title{addyrexp adds the years of experience by record to data.frame}
\usage{
addyrexp(indat)
}
\arguments{
\item{indat}{the docket catch-effort data.frame of interest}
}
\value{
the same input data.frame with the addition of one column
}
\description{
addyrexp adds the years of experience of each diver, as
    recorded in the docket catch-effort data.frame, to
    the right-hand side of the input data.frame. The years of
    experience for each diver will obviously change depending on
    which block, or collection of blocks, are selected, though it
    would be possible to apply this function to the whole data.frame
    of information. The value of this is that it facilates the
    exploration of the influence of years of experience on everything
    else.
}
\examples{
\dontrun{
  print("waiting on an integral data-set")
}
}
