% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{cbb}
\alias{cbb}
\title{cbb catch by block or smu plot}
\usage{
cbb(ind, smu = "block", legloc = "topright", catch = "catch", year = "year")
}
\arguments{
\item{ind}{the input data.frame or matrix}

\item{smu}{the column name of the smu variable default = "block"}

\item{legloc}{the location of the legend defaults to "topright",
could be "topleft", "bottomleft", or "bottomright"}

\item{catch}{the column name of the catc variable default = "catch"}

\item{year}{the column name of the year variable default = "year"}
}
\value{
invisibly the catch by year by block as a matrix
}
\description{
cbb plots out the catch by block or by smu, A legend is
    produced if there is a matrix of values and is made up of
    the column names of the input catchb matrix.
}
\examples{
\dontrun{
 ind=abd2;legloc="topright";smu="block";catch="catch";year="year"
}
}
