% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{cleanname}
\alias{cleanname}
\title{cleanname tidies a vector of names by removing commas and spaces}
\usage{
cleanname(invect, removeNA = TRUE)
}
\arguments{
\item{invect}{the vector of character strings to be tidied.}

\item{removeNA}{a boolean determining whether NA values are replaced with 
'unknown'}
}
\value{
a vector the same length as was input
}
\description{
cleanname tides character strings in an input vector by 
    removing commas, apostrophes, and spaces.
}
\examples{
x <- c("John Smith", "Port of Hobart","An example, and another")
cleanname(x)
}
