% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{coef.outce}
\alias{coef.outce}
\title{coef.outce S3 method to extract the parameter values from an outce object}
\usage{
\method{coef}{outce}(inout)
}
\arguments{
\item{inout}{is a outce object such as produced by standLM}
}
\value{
a matrix containing the optimum model parameters,
}
\description{
coef.outce extracts the parameter values from an outce
    object from a standardization (e.g. from standLM) by extending
    the generic 'coef' function to apply to outce objects. The output
    includes the log-space coefficients, the lower and upper 95%
    confidence intervals, the standard error of each parameter, and
    p the probability that the parameter is significantly different
    from zero. These coefficients would need to be back-transformed
    to return to the original scale. This would use
    x = exp(coef + (sterr*sterr)/2); note the bias correction
}
\examples{
\dontrun{
 data(sps)
 splabel = "Species"
 labelM <- c("Year","Vessel","Month")
 mods <- makemodels(labelM)
 sps1 <- makecategorical(labelM,sps)
 out <- standLM(mods,sps1,splabel)
 coef(out)
}
}
