% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{cvsh}
\alias{cvsh}
\title{cvsh plots the catch versus hours}
\usage{
cvsh(catchb, hoursb, legloc = "topleft")
}
\arguments{
\item{catchb}{a vector or matrix of catches by year by block}

\item{hoursb}{a matching vector or matrix of hours by year by block}

\item{legloc}{the location of the legend defaults to "topright",
could be "topleft", "bottomleft", or "bottomright", usually
determined interactively for each instance}
}
\value{
nothing, but it does plot a graph
}
\description{
cvsh plots the catch versus hours, one can obtain the
    catches and hours from the functions cbb and hbb. A legend is
    produced if there is a matrix of values and is made up of
    the column names of the input catchb matrix.
}
\examples{
\dontrun{
   ind=abd2;legloc="topright";sau="block";hours="hours";year="year"
   cbyb <- cbb(ind,sau)
   hbyb <- hbb(ind,sau)
   cvsh(cbyb,hbyb)
}
}
