% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{diagnosticPlot}
\alias{diagnosticPlot}
\title{diagnosticPlot produces diagnostic plots of the regression.}
\usage{
diagnosticPlot(
  inout,
  indat,
  inmodel = inout$Optimum,
  depend = "LnCE",
  inlabel = "",
  dorug = TRUE
)
}
\arguments{
\item{inout}{is the list from standLM.}

\item{indat}{is the data.frame used in the standardization.}

\item{inmodel}{is a single number identifying the model inside 'inout' to
plotted, defaults to inout's Optimum}

\item{depend}{the dependent variable defaults to "LnCE"}

\item{inlabel}{provides the means of giving a title to the 2 x 2 plot}

\item{dorug}{default=TRUE, should the rug of points be included in 
the QQplot?}
}
\value{
a 2 x 2 plot of the residuals, the Q-Q plot, and the distributions
   of the logged observed and predicted values of CPUE.
}
\description{
diagnosticPlot produces diagnostic plots of the regression.
  these include a plot of the residuals agsinst the predicted values, the
  normal Q-Q plot, a histogram of the Observed Log(CPUE) with a fitted
  normal distribution, and a histogram of the predicted Log(CPUE) with a
  fitted normal distribution and the normal distribution from the observed
  data.
}
\examples{
\dontrun{
data(sps)
splabel = "Species"
sps$Year <- factor(sps$Year)
sps$Month <- factor(sps$Month)
sps$Vessel <- factor(sps$Vessel)
labelM <- c("Year","Vessel","Month")
mods <- makemodels(labelM)
out <- standLM(mods,sps,splabel)
diagnosticPlot(out,3,sps,splabel)
}
}
