% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{divbb}
\alias{divbb}
\title{divbb plots a count of active diver by block by year}
\usage{
divbb(ind, smu = "block", legloc = "topright", diver = "diver", year = "year")
}
\arguments{
\item{ind}{input data.frame or matrix}

\item{smu}{the column name of the block variable default = "block"}

\item{legloc}{location of the legend defaults to "topright",
could be "topleft", "bottomleft", or "bottomright", usually
determined interactively for each instance}

\item{diver}{the column name of the diver variable default = "diver"}

\item{year}{the column name of the year variable default = "year"}
}
\value{
invisibly returns the count of active divers by year and block
}
\description{
divbb plots a count of divers active by block by year
}
\examples{
\dontrun{
  ind=abd2;legloc="topright";block="block";diver="diver";year="year"
}
}
