% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleanalysis.r
\name{dosingle}
\alias{dosingle}
\title{dosingle conducts a standardization of indat using the inmodel}
\usage{
dosingle(inmodel, indat)
}
\arguments{
\item{inmodel}{the formula used in the analysis; usually from the
function makeonemodel.}

\item{indat}{the data.frame containing the data to be analysed.}
}
\value{
a list with a similar structure to the out object, so not
   a outce class member but can be used with plotstand
}
\description{
dosingle conducts a standardization of indat using the
   inmodel.
}
