% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{examinedata}
\alias{examinedata}
\title{examinedata plots the yearly distributions of catch and effort from 'x'}
\usage{
examinedata(
  x,
  catch = "catch",
  labcatch = "catch",
  effort = "hours",
  labeffort = "effort",
  cpue = "cpue",
  labcpue = "cpue",
  year = "year",
  spsname = "",
  resdir = "",
  resfile = "",
  runname = "",
  plotnum = c(1, 1),
  wid = 6,
  hgt = 5,
  limitx = matrix(rep(NA, 6), nrow = 3, ncol = 2)
)
}
\arguments{
\item{x}{the data.frame of fishery dependent data}

\item{catch}{the name used to identify the catch factor of the species}

\item{labcatch}{the axis label for catch}

\item{effort}{the name used to identify the effort factor of the species}

\item{labeffort}{the axis label for effort}

\item{cpue}{the name used to identify the cpue factor of the species}

\item{labcpue}{the axis label for cpue}

\item{year}{the name used to identify the year factor of the species}

\item{spsname}{the name of the species of interest.}

\item{resdir}{the full path of the results directory into which the plot files
and the .csv files for the tables are to be stored.}

\item{resfile}{the full path and name of the '.csv' file used to store the 
filenames, category, and caption for each plot and table.}

\item{runname}{the name of the particular run being made}

\item{plotnum}{the number of rows an columns of plots used, default=c(1,1)}

\item{wid}{the width of each plot, default=6}

\item{hgt}{the height of each plot, default=5}

\item{limitx}{a matrix of 3 x 2 containing the xlim values for the first 
three graphs. Default is a 3 x 2 matrix of NA}
}
\value{
nothing but it does add 4 plots and two tables to the results
}
\description{
examinedata plots the yearly distributions of the number of 
    records, the catch, the effort, and the geometric mean cpue by year for 
    the data in the fishery dependent data in the data.frame x. It also 
    tabulates these values.  Each of these are stored either as .csv or .png 
    files 'in resdir'. It also logs these files in resfile, which opens the 
    possibility of displaying all results in a local webpage under a tab 
    labelled 'yeardata'. The 'resdir', 'resfile', and 'runname' should be the 
    same as those used by makehtml to prepare a directory and 'resfile' to 
    store the files to be used to make a local webpage of results.
}
\examples{
print("wait o internal data")
}
