% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{factorprop}
\alias{factorprop}
\title{factorprop summarizes the properties of a potential factor}
\usage{
factorprop(
  indat,
  fact,
  resdir,
  runname,
  resfile,
  year = "year",
  catch = "catch",
  effort = "hours",
  cpue = "cpue"
)
}
\arguments{
\item{indat}{the data.frame being used for the standardization}

\item{fact}{the name of the factor being explored, a character string}

\item{resdir}{the results directory into which all results for the runname are
put}

\item{runname}{the name of the particular analysis}

\item{resfile}{the result logfile, generated by setuphtml}

\item{year}{the name of the year factor in the data.frame. default='year'}

\item{catch}{the name of the catch factor in the data.frame. default='catch'}

\item{effort}{the name of the effort factor in the data.frame. default='hours'}

\item{cpue}{the name of the cpue factor in the data.frame. default='cpue'}
}
\value{
invisibly a list of records by year, catch by year and a table with
    total records, actively reporting years, and total catch for each value 
    of the factor
}
\description{
factorprop summarizes the properties of a potential factor to be
    used in a statistical standardization of cpue data. It does this in terms 
    of the number of records and catches by year, and how many years the 
    factor involved reported catches. Thus if dealing with divers and vessels
    it provides catch by diver or vessel by year, plus number of records
    reported, and summarizes all of these inside a local website. The value
    of this is that outliers and uneven distributions of records can be easily
    identified and decisions made about data selection and perhaps amalgamation
    of levels within a factor.
}
\examples{
print("need to wait on an internal data.frame")
}
