% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{filelist}
\alias{filelist}
\title{filelist literally lists the files in a directory}
\usage{
filelist(indir, findtext = "", ignorecase = FALSE, silent = FALSE)
}
\arguments{
\item{indir}{the directory whose contents are to be listed}

\item{findtext}{a text string to be searched for, default=""}

\item{ignorecase}{search on exact case or not, default=FALSE}

\item{silent}{should no summary be sent to the console, default=FALSE}
}
\value{
a dataframe of filenames is returned and, if silent=TRUE, two
    lines of information are sent to the console
}
\description{
filelist generate a data.frame of files in a directory
    including their properties - name, size, isdir, mdata and mtime,
    being the last modification data and time, adate and atime, being
    the last access dat and time. and finally a reference number to
    ease file selection. The option of searching for a text string
    within the complete filename is provided by the findtext argument.
}
\examples{
\dontrun{
  filelist(indir=getwd())
  filelist(indir=getwd(),findtext="txt",ignorecase=TRUE)
}
}
