% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{fishery}
\alias{fishery}
\title{fishery generates vectors of year, catch, effort, and cpue}
\usage{
fishery(
  indat,
  years = "year",
  catch = "catch",
  effort = "effort",
  cpue = "cpue"
)
}
\arguments{
\item{indat}{the input data.frame or matrix}

\item{years}{the name of the year factor; defaults to "year"}

\item{catch}{the name of the catch variable; defaults to "catch"}

\item{effort}{the name of effort variable; defaults to "effort"}

\item{cpue}{the name of the cpue variable; defaults to "cpue"}
}
\value{
a matrix with columns Year, Catch, Effort, Bias Corrected
   geometric mean CPUE, the naive geometric means, the arithemtic mean
   CPUE, and the number of records
}
\description{
fishery generates vectors of year, catch, effort, and cpue
   This permits a summary of the basic fishery characteristics within a
   data.frame or matrix. It requires the columns to be labelled.
}
\examples{
\dontrun{
data(sps)
fishery(sps,years="Year",catch="catch_kg",effort="Effort",cpue="CE")
}
}
