% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{geobb}
\alias{geobb}
\title{geobb plots the naive geometric mean by year and block}
\usage{
geobb(ind, smu = "block", legloc = "topright", LnCE = "LnCE", year = "year")
}
\arguments{
\item{ind}{the input data.frame or matrix}

\item{smu}{the column name of the block variable default = "block"}

\item{legloc}{the location of the legend defaults to "topright",
could be "topleft", "bottomleft", or "bottomright", usually
determined interactively for each instance}

\item{LnCE}{the column name of the LnCE variable default = "LnCE"}

\item{year}{the column name of the year variable default = "year"}
}
\value{
invisibly returns the naive gemetric mean of cpue by year and block
}
\description{
geobb plots the naive geometric mean by year and block.
    It expects to find a LnCE variable, the natural log of coue, in
    the input data.frame or matrix.
}
\examples{
\dontrun{
   print("wait")
}
}
