% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{getaav}
\alias{getaav}
\title{getaav calculates annual absolute variation in catch}
\usage{
getaav(invect, narm = TRUE)
}
\arguments{
\item{invect}{a vector of catches}

\item{narm}{boolean, should NAs be removed? If not then getaav will
return an NA.  Default = TRUE}
}
\value{
a single scalar value the AAV of the input catches
}
\description{
getaav calculates the annual absolute change in catch
    for an input vector of catches, which could be across a series
    of years or even across different spatial units for a single
    year (an unusual use).
    The equation used is aav = 100 x sum(|Ct - Ct-1|)/(sum(Ct).
}
\examples{
  catch <- c(1,2,3,4,5,4,3,2,1)
  getaav(catch)  # should equal 0.32
}
