% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{getfact}
\alias{getfact}
\title{getfact extracts parameter estimates for a given factor}
\usage{
getfact(inmat, invar, biascorrect = TRUE)
}
\arguments{
\item{inmat}{this can be an outce object from standLM, but it can also be
a matrix of coefficients, an lm object, or a gam object}

\item{invar}{the model variable whose parameters are wanted, eg "month"}

\item{biascorrect}{when back transforming the coefficients should we use the
log-normal bias-correction or not. default=TRUE}
}
\value{
a matrix containing the parameters for invar
}
\description{
getfact extracts the parameter estimates for a given factor
    from either a matrix, an array, an outce object (from standLM), or an 
    lm object or a gam object. It does this by searching the rownames of the
    output parameters of the optimum model. It also checks for interaction
    terms, which for categorical factors is the same as determining a
    trend of the two factors relative to each other. e.g. for Zone:Month
    the outcome is the monthly trend for each zone.
}
