% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{getStand}
\alias{getStand}
\title{getStand extracts major results from an outce object}
\usage{
getStand(x, P = 0.95)
}
\arguments{
\item{x}{the outce object derived from standLM}

\item{P}{the probability interval for the confidence intervals}
}
\value{
matrix containing the primary results
}
\description{
getStand extracts major results from an outce object. which
   is generated using standLM. The main results include the year, the
   geometric mean (the first column of the standardization), the optimum
   model, the standard error of the optimum, and the lower and upper
   confidence intervals
}
\examples{
\dontrun{
 data(abeg)
 splabel = "Mollusc"
 labelM <- c("year","diverID","month","block","boatID")
 ab1 <- makecategorical(labelM,ab)
 mods <- makemodels(labelM)
 out <- standLM(mods,ab1,splabel)
 round(out$Results,4)
 round(out$WhichM,4)
 getStand(out)
}
}
