% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{hbb}
\alias{hbb}
\title{hbb plots the hours by year by block}
\usage{
hbb(ind, smu = "block", legloc = "topright", hours = "hours", year = "year")
}
\arguments{
\item{ind}{the input data.frame or matrix}

\item{smu}{the column name of the block variable default = "block"}

\item{legloc}{the location of the legend defaults to "topright",
could be "topleft", "bottomleft", or "bottomright"}

\item{hours}{the column name of the effort variable default = "hours"}

\item{year}{the column name of the year variable default = "year"}
}
\value{
invisibly a matrix of effort(hours) by year by block
}
\description{
hbb plots the hours by year by block
}
\examples{
\dontrun{
 ind=abd2;legloc="topright";block="block";hours="hours";year="year"

}
}
