% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{makeLabel}
\alias{makeLabel}
\title{makeLabel Convert a vector of numbers or strings into a single label}
\usage{
makeLabel(invect, insep = "_")
}
\arguments{
\item{invect}{the vector of numbers or strings to be converted into a
single string.}

\item{insep}{defaults to '_' but can be any selected character used to
separate each value}
}
\value{
a character string containing the invect as a single string.
}
\description{
makeLabel Convert a vector of numbers of strings into a single
  label
}
\examples{
\dontrun{
 x <- c(1,2,3,4,5)
 makeLabel(x)
 makeLabel(x,"-")
}
}
