% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleanalysis.r
\name{makeonemodel}
\alias{makeonemodel}
\title{makeonemodel generates a single model for use in dosingle}
\usage{
makeonemodel(labelModel, dependent = "LnCE")
}
\arguments{
\item{labelModel}{a vector of labels for each factor to be included
in the analysis}

\item{dependent}{the name of the dependent variable; defaults to
LnCE}
}
\value{
a formula of 'dependent ~ labelModel components
}
\description{
makeonemodel puts together the formula needed to
    run a statistical model, but, different to makemodels, it
    only generates a single model.
}
\examples{
labelM <- c("Year","Vessel","DepCat","Zone:Month")
makeonemodel(labelM)
makeonemodel(labelM,dependent="LnCE")
}
