% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{maps}
\alias{maps}
\title{maps - lists the important functions with their syntax ready for use}
\usage{
maps()
}
\value{
prints the useable functions and their syntax ready for use.
}
\description{
maps - lists the important functions with their syntax ready
  for use. Just copy the printed text and modify to suit your own purposes.
}
\examples{
maps()
}
