% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{oneyrdivers}
\alias{oneyrdivers}
\title{oneyrdivers removes divers who only fished for one year}
\usage{
oneyrdivers(indat)
}
\arguments{
\item{indat}{the data.frame of abalone fishery data}
}
\value{
a data.frame with single year divers removed
}
\description{
oneyrdivers takes in a data.frame of abalone fishery
    data and removes records relating to divers who only fished for
    a total of one year. Some only have a few records but there are
    instances of some divers having hundreds of records. In most
    standardizations single year divers have no influence, although
    one would expect an influence if one treated divers as a random
    effect.
}
\examples{
\dontrun{
 print("still being developed")
}
}
