% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{plotaus}
\alias{plotaus}
\title{plotaus sets up a schematic map outline of the SESSF}
\usage{
plotaus(
  leftlong = 129,
  rightlong = 155,
  uplat = -25,
  downlat = -44.6,
  defineplot = T,
  gridon = 0,
  maintitle = "",
  infont = 7
)
}
\arguments{
\item{leftlong}{the longitude of the left hand side of the box plotted;
defaults to 129.0}

\item{rightlong}{the longitude of the right hand size of the box plotted;
defaults to 155.0}

\item{uplat}{the latitude (negative) of the top of the box; default -25.0}

\item{downlat}{nagative latitude of the bottom of the box; default -44.6}

\item{defineplot}{default = TRUE; declare the par values for a plot. If this
is set to FALSE it would be possible to plot out multiple maps in one plot.}

\item{gridon}{plot lines at the value of 'gridon'. If set to 0.5 it would
include lines every half degree.}

\item{maintitle}{default = "", adds a title to the plot at top center.}

\item{infont}{default = 7, bold Times. Used to define the font for axes, etc}
}
\value{
plots a schematic map of the SESSF, If defineplot=TRUE it will alter
  the default par values for plots.
}
\description{
plotaus sets up a schematic map outline of Australia. Only the
    landmass contained within a box defined by leftlong, rightlong, uplat,
    and downlat. is included. The land areas is shaded.
}
\examples{
\dontrun{
 dev.new(height=6.0,width=7.5,noRStudioGD = TRUE)
 plotaus()
}
}
