% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{plotdata}
\alias{plotdata}
\title{plotdata generates graphs of CE and log-transformed CE data.}
\usage{
plotdata(indat, dependent = "LnCE")
}
\arguments{
\item{indat}{is the data.frame containing the data being standardized.}

\item{dependent}{variable name; defaults to LnCE.}
}
\value{
a plot of the untransformed cpue data and of the log-transformed
  data. Included on the log-transformed data is a fitted normal
  distribution and the bias-corrected mean estiamte of average CPUE.
}
\description{
plotdata generates graphs of CE and log-transformed CE
  data. Included in the graph of the log-transformed cpue data is a
  fitted normal distribution with the associated bias-corrected average.
}
\examples{
\dontrun{
data(sps)
pick <- which(sps$Year == 2005)
sps2 <- droplevels(sps[pick,])
plotprep()
plotdata(sps2)
plotdata(sps2[sps2$Depth > 200,])
}
}
