% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{plotpoly}
\alias{plotpoly}
\title{plotpoly adds a 'gridon' box to the map}
\usage{
plotpoly(inlong, inlat, col = 1, gridon = 0.1)
}
\arguments{
\item{inlong}{top left longitude}

\item{inlat}{top left latitude}

\item{col}{the colour of the box polygon, defaults to black}

\item{gridon}{the size of the box, defaults to 0.1 degree}
}
\value{
nothing, the function adds a rectangular polygon to map
    it is not exported
}
\description{
plotpoly adds a 'gridon' box to the map t is purely
    used within plotpolys and so is not exported
}
