% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{plotpolys}
\alias{plotpolys}
\title{plotpolys - plots a grid and fills it with relative catches}
\usage{
plotpolys(
  intmp,
  leftlong,
  rightlong,
  uplat,
  downlat,
  leg = "topleft",
  mincount = 1,
  intitle = " ",
  gridon = 0.1,
  hot = F,
  map = T,
  scaler = NA,
  textout = TRUE,
  namecatch = "catch_kg"
)
}
\arguments{
\item{intmp}{- a matrix or data.frame containing, at least, columns with the
names 'Long', 'Lat', and 'catch_kg'.}

\item{leftlong}{the longitude of the left hand side of the box plotted}

\item{rightlong}{the longitude of the right hand size of the box plotted}

\item{uplat}{the latitude (negative) of the top of the box}

\item{downlat}{nagative latitude of the bottom of the box}

\item{leg}{- legend location, default is 'topleft', alternatives are
'topright', 'center', 'left', and 'right'}

\item{mincount}{- the minimum number of observations in a grod cell before it
is included in the colour plotting; default - 1.0}

\item{intitle}{- default="". Provides for a title top center.}

\item{gridon}{- default 0.1 = 0.1 degree square. could be 0.2, 0.25, 0.5, etc}

\item{hot}{- default FALSE. If true then instead of 8 levels of catch
intensity there are only 4 and these differ each by an order of magnitude.}

\item{map}{- default TRUE, use FALSE is only the tabulation of catch data into grid cells
is required without a plot.}

\item{scaler}{- default to NA, otherwise used to select which scale to use
valid values are 1 - 9. This is used to set the same scale in each of
multiple plots.}

\item{textout}{default = TRUE; determines whether the summary information
from the tabulation is pronted to screen or not}

\item{namecatch}{the name of the variable which to sum into each polygon.
this defaults to 'catch_kg', but could be 'Effort', etc.}
}
\value{
invisibly returns a matrix of grids with the total catch in each cell
   in units of tonnes. Also, if map = TRUE, adds the grid and a colour rendition
   to the map.
}
\description{
plotpolys - uses the same input object as plotaus plus the
   leftlong, rightlong, uplat, and downlat variables to which it adds a grid
   defined by 'gridon'. It then sums the catches found in each grid and then
   plots this as a patchwork of colours representing the relative density.
   Uses the function 'plotpoly'.
}
\examples{
\dontrun{
plotaus()
}
}
