% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{plotstand}
\alias{plotstand}
\title{plotstand plot optimum model from standLM  vs Geometric mean}
\usage{
plotstand(stnd, bars = FALSE, geo = 0, P = 95, catch = NA, usefont = 7)
}
\arguments{
\item{stnd}{is the list output from standLM}

\item{bars}{is a logical T or F determining whether to put confidence bounds
around each estimate; defaults to FALSE}

\item{geo}{is an estimate of the original geometric mean catch rate across
all years. If this is > 0.0 it is used to rescale the graph to the
nominal scale, otherwise the mean of each time-series will be 1.0, which
simplifies visual comparisons. geo defaults to 0.0.}

\item{P}{is the percentile used for the log-normal confidence bounds, if
they are plotted; defaults to 95.}

\item{catch}{if it is desired to plot the catch as well as the CPUE
then a vector of catches needs to be input here}

\item{usefont}{enables the font used in the plot to be modified. Most
publications appear to prefer usefont=1; defaults to 7 - Times bold}
}
\value{
a plot of the model with the smallest AIC (solid line) and the
  geometric mean (model 1, always = LnCE ~ Year, the dashed line). 'Year'
  could be some other time step.
}
\description{
plot optimum model from standLM  vs Geometric mean.
  Has options that allow for log-normls P% intervals around each time
  period's parameter estimate. Also can rescale the graph to have an average
  the same as the geometric mean average of the original time series of data.
}
\examples{
\dontrun{
data(abeg)
splabel = "SpeciesName"
labelM <- c("year","diver","month")
ab1 <- makecategorical(labelM[1:3],abeg)
mods <- makemodels(labelM)
out <- standLM(mods,ab1,splabel)
plotprep()
plotstand(out, bars=TRUE, P=90,geo=100.0,usefont=1)
plotstand(out)
}
}
