% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{plotyrexp}
\alias{plotyrexp}
\title{plotyrexp generates a plot of count by yrs of experience diving}
\usage{
plotyrexp(indat, title = "")
}
\arguments{
\item{indat}{the data.frame of docket abalone catch-effort fishery data}

\item{title}{a title for the plot, defaults top ""}
}
\value{
invisibly returns the counts of divers by years of experience
}
\description{
plotyrexp generates a plot of count of divers versus
    years of experience from teh input data.frame. Exeperience is
    defined as the number of years in which the diver reports
    catches in the docket catch-effort data.frame.
}
\examples{
\dontrun{
  print("waiting on an inbuilt data-set")
}
}
