% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{qqdiag}
\alias{qqdiag}
\title{qqdiag generates a qqplot with a histogram of residuals}
\usage{
qqdiag(
  inmodel,
  plotrug = FALSE,
  bins = NA,
  hline = 0,
  xinc = 100,
  yinc = 1.1,
  ylab = "residuals"
)
}
\arguments{
\item{inmodel}{the optimum model being considered}

\item{plotrug}{a logical term determining whether a rug is plotted on the
qqplot.}

\item{bins}{defaults to NA, but can be set to a given series}

\item{hline}{Include some horizontal lines on the histogram. defaults to 0.}

\item{xinc}{the increment for tick marks on the xaxis of the histogram}

\item{yinc}{the increment for tick marks on the y-axis of the histogram}

\item{ylab}{the y-axis label for the histogram, defaults to 'residuals'}
}
\value{
plots a graph and invisibly returns the output from the histogram
}
\description{
qqdiag generates a qqplot with a complementary histogram of
    the residuals to illustrate the proportion of all residuals along the
    qqline. If the qqline deviates from the expected straigt line, which
    is red i colour to make for simpler comparisons, then the histogram
    enables one to estiamte what proportion of records deviate from
    normality. The zero point is identified with a line, as are the
    approximate 5% and 95% percentiles. In both cases > 5% is above or
    below the blue lines, with < 90% in between depending on the
    proportions in each class. To get a more precise estimate use the
    invisibly returned histogram values.
}
\examples{
\dontrun{
 y <- rep(1:100,2)
 x <- rnorm(200,mean=10,sd=1)
 model <- lm(y ~ x)
 dev.new(width=6,height=3.5,noRStudioGD = TRUE)
 par(mai=c(0.45,0.45,0.15,0.05),font.axis=7)
 qqdiag(model,xinc=1,yinc=10,bins=seq(-55,50,2.5))
}
}
