% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{qqplotout}
\alias{qqplotout}
\title{qqplotout plots up a single qqplot for a lm model}
\usage{
qqplotout(
  inmodel,
  title = "Normal Q-Q Plot",
  cex = 0.9,
  ylow = -5,
  yhigh = 5,
  plotrug = FALSE
)
}
\arguments{
\item{inmodel}{the optimum model from standLM or dosingle}

\item{title}{a title for the plot, defaults to 'Normal Q-Q Plot'}

\item{cex}{the size of the font used, defaults to 0.9}

\item{ylow}{the lower limit of the residuals}

\item{yhigh}{he upper limit of the residuals}

\item{plotrug}{a logical value determinning whether a rug is included}
}
\value{
currently nothing, but it does generate a qqplot to the current
    device
}
\description{
qqplotout generates a single qqplot in isolation from the
    plot of a model's diagnostics. It is used with lefthist to
    illustrate how well a model matches a normal distribution
}
\examples{
\dontrun{
 y <- rep(1:100,2)
 x <- rnorm(200,mean=10,sd=1)
 model <- lm(y ~ x)
 dev.new(width=6,height=3.5,noRStudioGD = TRUE)
 par(mai=c(0.45,0.45,0.15,0.05),font.axis=7)
 qqplotout(model,ylow=-50,yhigh=50)
}
}
