% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{removeEmpty}
\alias{removeEmpty}
\title{removeEmpty - removes empty strings from a vector of strings}
\usage{
removeEmpty(invect)
}
\arguments{
\item{invect}{- a vector of characters}
}
\value{
a vector of strings without empty ones
}
\description{
removeEmpty - removes empty strings from a vector of strings
}
\examples{
inputtxt <- "a  b   c   g"
x <- unlist(strsplit(inputtxt," "))
print(x)
print(removeEmpty(x))
}
