% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{saveresults}
\alias{saveresults}
\title{saveresults sends the tables generated to a text file in resultdir}
\usage{
saveresults(sps, answer, out, mods, splabel, rundir)
}
\arguments{
\item{sps}{the data.frame containing all data relating to the species}

\item{answer}{the list obtained from 'selectdata' containing the data.frame
of the selected data, and the nobs table containing how different
selection criteria affect the number of observations.}

\item{out}{the output from standLM or dosingle containing the
standardization}

\item{mods}{the mods defined for the standardization}

\item{splabel}{the name of the species and fishery}

\item{rundir}{the path to the results directory}
}
\value{
nothing, but it does write a text file to the results directory.
}
\description{
savereults saves all the tables generated to a text file
    named after the species and fishery and puts it into resultdir. The
    outputs include for all data relating to the species: catch-by-method,
    catch-by-zone, catch-by-fishery, then only for the data selected for
    the standardization, the: catch-by-method, catch-by-zone, and the
    catch-by-DepCat, but also the number of observations modified by each
    of the selction criteria, the data summary, and the raw output from the
    standardizatio.
}
\examples{
\dontrun{
print("this will take some doing")
}
}
