% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{scaleCE}
\alias{scaleCE}
\title{scaleCE scales an input vector of CPUE to a mean of one x avCE}
\usage{
scaleCE(invect, avCE = 1)
}
\arguments{
\item{invect}{a vector of linear scale CPUE}

\item{avCE}{defaults to one but can be set to any particular value}
}
\value{
a vector of CPUE re-scaled to a mean of one or avCE
}
\description{
scaleCE scales a vector of CPUE to a mean of
    one or avCE. The use of a mean of one means that visual comparisons
    between different time-series becomes visually simplified. The
    avCE option could be used to scale the CPUE to the average
    geometric mean - so as to put it on the nominal scale
}
\examples{
\dontrun{
 ce <- c(0.4667187,1.2628564,0.8442146,0.9813531, 0.5554076,0.7426321)
 scaleCE(ce)
 scaleCE(ce,100.0)
}
}
