% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{tapsum}
\alias{tapsum}
\title{tapsum simplifies the use of tapply for summarizing variables}
\usage{
tapsum(indat, first, second, third = NA, div = 1000)
}
\arguments{
\item{indat}{the data.frame containing the raw fishery data}

\item{first}{the variable name (in quotes) being summed}

\item{second}{the first grouping variable}

\item{third}{the second grouping variable, defaults to NA}

\item{div}{defaults to 1000 to change Kg to tonnes. set t0 1.0 or NA to
avoid its influence}
}
\value{
a vector or matrix of sums of the pickvar by the frist and optionally
     the second grouping variable
}
\description{
data exploration commonly uses the tapply function and tapsum
    simplifies its use when obtaining the sum of any variable relative to
    other variables. For example it is common to want the total catch by
    year and, for example, Month, DepCat, Zone, etc.
}
\examples{
\dontrun{
  data(sps)
  tapsum(sps,"catch_kg","Year","Month")
}
}
