% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{xyplotyear}
\alias{xyplotyear}
\title{xyplotyear generates nyear xy plots from a data.frame}
\usage{
xyplotyear(
  x,
  yvar = "",
  xvar = "",
  year = "year",
  plotnum = c(1, 1),
  xlim = c(0, 12),
  addline = TRUE,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{x}{the data.frame containing the data}

\item{yvar}{the character name of y-axis column in the data.frame x}

\item{xvar}{the character name of x-axis column in the data.frame x}

\item{year}{the name of the year variable, default="year"}

\item{plotnum}{a vector of rows and cols for the plots, default=c(1,1). This
assumes that the columns are filled first using mfcol}

\item{xlim}{the range of the xvar to be plotted, default=c(0,12)}

\item{addline}{should a linear regression be fitted and added to each plot
default=TRUE}

\item{xlab}{the  generic label for the x-axis, default='', if left empty the 
xvar name will be used}

\item{ylab}{the generic label for the y-axis, default='', if left empty the 
yvar name will be used}
}
\value{
currently nothing but it does plot a graph
}
\description{
xyplotyear meets a common need that occurs when we have xy data
    from multiple years and want to plot them we can use xyplotyear. The 
    y-label for each plot is the year of data. The numeric label at the top 
    of each plot includes the number of observations, the gradient of the
    regression, if included, and the sum of the yvar for each year. The same
    y-axis scale is used for each plot.
}
\examples{
print("wait on internal data")
}
