% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{yearNA}
\alias{yearNA}
\title{yearNA - counts NAs per year in each numeric field in a data.frame}
\usage{
yearNA(indat, years = "Year", empty = FALSE)
}
\arguments{
\item{indat}{the data.frame whose numeric fields are to be considered}

\item{years}{identifies the name of the "Year" field in the data.frame}

\item{empty}{logical default=FALSE, determines whether columns which have
no NA present are printed or not}
}
\value{
a matrix of years x numeric fields with the number of records per
   field per year as reference
}
\description{
yearNA - counts the number of NAs in each year of each numeric
   field in a data.frame and outputs the results as a matrix
}
\examples{
\dontrun{
 year <- sort(rep(1990:1994,5))
 columns <- c("year","Var1","Var2","Var3")
 dat <- matrix(runif(100),nrow=25,ncol=4,dimnames=list(year,columns))
 dat[trunc(100*runif(20))] <- NA
 dat[,1] <- year
 print(dat)
 yearNA(as.data.frame(dat),years="year")
}
}
