% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_functions.r
\name{yearprop}
\alias{yearprop}
\title{yearprop summarizes the properties of the year as a factor}
\usage{
yearprop(
  indat,
  resdir,
  runname,
  resfile,
  year = "year",
  catch = "catch",
  effort = "hours",
  cpue = "cpue"
)
}
\arguments{
\item{indat}{the data.frame being analysed}

\item{resdir}{the results directory into which all results for the runname are
put}

\item{runname}{the name of the particular analysis}

\item{resfile}{the result logfile, generated by setuphtml}

\item{year}{the name of the year factor in the data.frame. default='year'}

\item{catch}{the name of the catch factor in the data.frame. default='catch'}

\item{effort}{the name of the effort factor in the data.frame. default='hours'}

\item{cpue}{the name of the cpue factor in the data.frame. default='cpue'}
}
\value{
invisibly returns a matrix of year, records, catch, hours, and 
   bias-corrected geometric mean cpue. Also adds a table to the resdir and 
   resfile
}
\description{
yearprop summarizes the properties of the year as a factor,
    which is useful before conducting a cpue standardization. It also outputs
    a table of the properties of the input data.frame
}
\examples{
print("need to wait on an internal data.frame")
}
