% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{histyear}
\alias{histyear}
\title{histyear plots a histogram of a given variable for each year available}
\usage{
histyear(
  x,
  Lbound = -3.5,
  Rbound = 12.25,
  inc = 0.25,
  pickvar = "LnCE",
  years = "Year",
  varlabel = "log(CPUE)",
  vline = NA,
  plots = c(3, 3)
)
}
\arguments{
\item{x}{the data.frame of data with at least a 'Year' and pickvar present}

\item{Lbound}{leftbound on all histograms, defaults to -3.5}

\item{Rbound}{right bound on all histograms, defaults to 12.25}

\item{inc}{the class width of the histogram, defaults to 0.25}

\item{pickvar}{which variable to plot each year default = 'LnCE'}

\item{years}{which variable name identifies the yaer column, default='Year'}

\item{varlabel}{what label to use on x-axis, default = 'log(CPUE)'}

\item{vline}{an optional vertical line to aid interpretation. If it is
numeric it will be added to each plot}

\item{plots}{how many plots to generate, default = c(3,3)}
}
\value{
a matrix of the year, mean value, stdev, and N number of
    observations. It also plots a histogram for each year and fits a
    normal distribution to each one.
}
\description{
histyear plots a histogram of a given variable for each year
    available
}
\examples{
\dontrun{
print("still to be developed")
}
}
