% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{plotbasic}
\alias{plotbasic}
\title{plotbasic generates 6 plots of fisheries data and diver data}
\usage{
plotbasic(
  indat,
  title,
  sau = "block",
  leg1 = "topleft",
  leg2 = "bottomleft",
  leg3 = "bottomleft"
)
}
\arguments{
\item{indat}{the data.frame containing the fisheries data}

\item{title}{the main title for the plots}

\item{sau}{the spatial management unit used, in Tasmania = block}

\item{leg1}{legend location for the catch vs hours by sau plot,
default="topleft"}

\item{leg2}{legend location for the geometric mean cpue by sau plot,
default="bottomleft"}

\item{leg3}{legend location for the active diver by sau plot,
default="bottomleft"}
}
\value{
nothing but it generates a 3 row x 2 col plot
}
\description{
plotbasic expects to receive plots six graphs,
}
\examples{
\dontrun{
  indat=abd; title="Fortescue";sau="block";leg1="topleft";
  leg2="bottomleft";leg3="bottomleft"
  print("Waiting on an internal data.frame")
}    # indat=abd; title="blk13E"; sau="block";
}
